==> analyses/oncoprint-landscape/02-plot-oncoprint.R <==
# This script displays an oncoprint displaying the landscape across PBTA given
# the relevant metadata. It addresses issue #6 in the OpenPBTA-analysis
# github repository. It uses the output of 00-map-to-sample_id.R. It can
# accept a gene list file or a comma-separated set of gene list files that will
# be concatenated and restricts plotting to those genes (via --goi_list).
#
# Code adapted from the PPTC PDX Oncoprint Generation repository here:
# https://github.com/marislab/create-pptc-pdx-oncoprints/tree/master/R
#
# Chante Bethell for CCDL 2019 and Jo Lynne Rokita
#
# EXAMPLE USAGE:
#
# Rscript --vanilla 02-plot-oncoprint.R \
#   --maf_file ../../scratch/oncoprint_files/primary_only_maf.tsv \
#   --cnv_file ../../scratch/oncoprint_files/primary_only_cnv.tsv \
#   --fusion_file ../../scratch/oncoprint_files/primary_only_fusions.tsv \
#   --metadata_file ../../data/pbta-histologies.tsv \
#   --broad_histology "Embryonal tumor" \
#   --goi_list data/embryonal-tumor_goi_list.tsv \
#   --top_n 20 \
#   --png_name plots/primary_only_embryonal_oncoprint.png \
#   --output_table primary_only_embryonal_oncoprint_summary_n.tsv


#### Set Up --------------------------------------------------------------------

# Load maftools
library(maftools)

# Get `magrittr` pipe
`%>%` <- dplyr::`%>%`

#### Directories and Files -----------------------------------------------------

# Detect the ".git" folder -- this will in the project root directory.
# Use this as the root directory to ensure proper execution, no matter where
# it is called from.
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))

# Path to output directory for plots produced
plots_dir <-
  file.path(root_dir, "analyses", "oncoprint-landscape", "plots")

if (!dir.exists(plots_dir)) {
  dir.create(plots_dir)
}

# Path to output directory for tables produced
tables_dir <-
  file.path(root_dir, "analyses", "oncoprint-landscape", "tables")

if (!dir.exists(tables_dir)) {
  dir.create(tables_dir)
}

# Source the custom functions script
source(
  file.path(
    root_dir,
    "analyses",
    "oncoprint-landscape",
    "util",
    "oncoplot-functions.R"
  )
)

#### Command line options ------------------------------------------------------

# Declare command line options
option_list <- list(
  optparse::make_option(
    c("-m", "--maf_file"),
    type = "character",
    default = NULL,
    help = "file path to MAF file that contains SNV information",
  ),
  optparse::make_option(
    c("-c", "--cnv_file"),
    type = "character",
    default = NULL,
    help = "file path to file that contains CNV information"
  ),
  optparse::make_option(
    c("-f", "--fusion_file"),
    type = "character",
    default = NULL,
    help = "file path to file that contains fusion information"
  ),
  optparse::make_option(
    c("-s", "--metadata_file"),
    type = "character",
    default = NULL,
    help = "file path to the histologies file"
  ),
  optparse::make_option(
    c("-g", "--goi_list"),
    type = "character",
    default = NULL,
    help = "comma-separated list of genes of interest files that contain the
            genes to include on oncoprint"
  ),
  optparse::make_option(
    c("-b", "--broad_histology"),
    type = "character",
    default = NULL,
    help = "optional name of `broad_histology` value to plot associated oncoprint"
  ),
  optparse::make_option(
    c("-n", "--top_n"),
    type = "integer",
    default = 25,
    help = "`n` to display top n genes based on count of mutations, default is 25"
  ),
  optparse::make_option(
    c("-p", "--png_name"),
    type = "character",
    default = NULL,
    help = "oncoprint output png file name"
  ),
  optparse::make_option(
    c("--include_introns"),
    action = "store_true",
    default = FALSE,
    help = "logical statement on whether to include intronic variants in oncoprint plot"
  ),
  optparse::make_option(
    c("--output_table"),
    type = "character",
    default = NULL,
    help = "file name for the output tables"
  )
)

# Read the arguments passed
opt_parser <- optparse::OptionParser(option_list = option_list)
opt <- optparse::parse_args(opt_parser)

# Define cnv_file, fusion_file, and genes object here as they still need to
# be defined for the `prepare_and_plot_oncoprint` custom function (the
# cnv_file specifically for the `read.maf` function within the custom function),
# even if they are NULL
cnv_df <- opt$cnv_file
fusion_df <- opt$fusion_file
goi_list <- opt$goi_list

#### Read in data --------------------------------------------------------------

# Read in metadata
metadata <- readr::read_tsv(opt$metadata_file, guess_max = 10000) %>%
  dplyr::rename(Tumor_Sample_Barcode = sample_id)

# Read in MAF file
maf_df <- data.table::fread(opt$maf_file,
                            stringsAsFactors = FALSE,
                            data.table = FALSE)

if (!opt$include_introns) {
  maf_df <- maf_df %>%
    dplyr::filter(Variant_Classification != "Intron")
}

# Read in cnv file
if (!is.null(opt$cnv_file)) {
  cnv_df <- readr::read_tsv(opt$cnv_file) 
}

# Read in fusion file and join
if (!is.null(opt$fusion_file)) {
  fusion_df <- readr::read_tsv(opt$fusion_file)
}

#### Set up oncoprint annotation objects --------------------------------------
# Read in histology standard color palette for project
histology_label_mapping <- readr::read_tsv(
  file.path(root_dir,
            "figures",
            "palettes", 
            "broad_histology_cancer_group_palette.tsv")) %>%
  # Select just the columns we will need for plotting
  dplyr::select(broad_histology, broad_histology_display, cancer_group, cancer_group_display, cancer_group_hex)

# Join on these columns to the metadata
metadata <- metadata %>% 
  dplyr::inner_join(histology_label_mapping, by = c("broad_histology", "cancer_group"))

# Filter to the metadata associated with the broad histology value, if provided
if (!is.null(opt$broad_histology)) {
  
  if (!opt$broad_histology == "Other CNS") {
    metadata <- metadata %>%
      dplyr::filter(broad_histology == opt$broad_histology)
  } else {
    metadata <- metadata %>%
      dplyr::filter(
        broad_histology %in% c(
          "Ependymal tumor",
          "Tumors of sellar region",
          "Neuronal and mixed neuronal-glial tumor",
          "Tumor of cranial and paraspinal nerves",
          "Meningioma",
          "Mesenchymal non-meningothelial tumor",
          "Germ cell tumor",
          "Choroid plexus tumor",
          "Histiocytic tumor",
          "Tumor of pineal region",
          "Metastatic tumors",
          "Other astrocytic tumor",
          "Lymphoma",
          "Melanocytic tumor",
          "Other tumor"
        )
      )
  }
  
  # Now filter the remaining data files
  maf_df <- maf_df %>%
    dplyr::filter(Tumor_Sample_Barcode %in% metadata$Tumor_Sample_Barcode)
  
  cnv_df <- cnv_df %>%
    dplyr::filter(Tumor_Sample_Barcode %in% metadata$Tumor_Sample_Barcode)
  
  fusion_df <- fusion_df %>%
    dplyr::filter(Tumor_Sample_Barcode %in% metadata$Tumor_Sample_Barcode)
  
}

# Read in the oncoprint color palette
oncoprint_col_palette <- readr::read_tsv(file.path(
  root_dir,
  "figures",
  "palettes",
  "oncoprint_color_palette.tsv"
)) %>%
  # Use deframe so we can use it as a recoding list
  tibble::deframe()

# Color coding for `cancer_group` classification
# Get unique tumor descriptor categories
histologies_color_key_df <- metadata %>%
  # arrange by cancer_group_display *counts*
  dplyr::mutate(cancer_group_display = forcats::fct_infreq(cancer_group_display)) %>%
  dplyr::arrange(cancer_group_display) %>%
  dplyr::select(cancer_group_display, cancer_group_hex) %>%
  dplyr::distinct()

# Make color key specific to these samples
histologies_color_key <- histologies_color_key_df$cancer_group_hex
names(histologies_color_key) <- histologies_color_key_df$cancer_group_display

# Now format the color key objet into a list
annotation_colors <- list(cancer_group_display = histologies_color_key,
                          germline_sex_estimate = c("Male"="#2166ac",
                                                    "Female"="#b2182b"))

#### Prepare MAF object for plotting ------------------------------------------

maf_object <- prepare_maf_object(
  maf_df = maf_df,
  cnv_df = cnv_df,
  metadata = metadata,
  fusion_df = fusion_df
)


#### Subset MAF Object (Optional)----------------------------------------------

# Code here is specifically adapted from:
# https://github.com/marislab/create-pptc-pdx-oncoprints/blob/master/R/create-complexheat-oncoprint-revision.R

# We only need to subset the GOI list if there are more GOI than the top n argument
# Subset `maf_object` for histology-specific goi list
if (!is.null(opt$goi_list)){
  
  # Read in genes of interest information using the `read_tsv()` function
  goi_list <- readr::read_tsv(opt$goi_list) %>%
    as.matrix()
  
  filtered_maf_object <- subsetMaf(
    maf = maf_object,
    tsb = metadata$Tumor_Sample_Barcode,
    genes = goi_list,
    mafObj = TRUE
  )
  
  # Get top mutated genes per this subset object
  gene_sum <- mafSummary(filtered_maf_object)$gene.summary
  
  # write out results 
  if (!is.null(opt$output_table)) {
    readr::write_tsv(gene_sum, file.path("tables", opt$output_table))
  }
  
  # Sort to get top altered genes rather than mutated only genes
  goi_list <- gene_sum %>%
    dplyr::arrange(dplyr::desc(AlteredSamples)) %>%
    # Filter to genes where multiple samples have an alteration
    dplyr::filter(AlteredSamples > 1) %>%
    dplyr::pull(Hugo_Symbol)

  if (opt$top_n < length(goi_list)) {
    # Now let's filter to the `top_n` genes
    goi_list <- goi_list[1:opt$top_n]

  }
}

### Plot and Save Oncoprint --------------------------------------------------

# Given a maf object, plot an oncoprint of the variants in the
# dataset and save as a png file.
png(
  file.path(plots_dir, tolower(gsub(" ", "-", opt$png_name))),
  width = 35,
  height = 20,
  units = "cm",
  res = 300
)

oncoplot(
  maf_object,
  clinicalFeatures = c("cancer_group_display","germline_sex_estimate"),
  genes = goi_list,
  logColBar = TRUE,
  sortByAnnotation = TRUE,
  showTumorSampleBarcodes = FALSE,
  removeNonMutated = TRUE,
  annotationFontSize = 1.25,
  SampleNamefontSize = 1,
  fontSize = 1,
  colors = oncoprint_col_palette,
  annotationColor = annotation_colors,
  bgCol = "#F5F5F5",
  top = opt$top_n,
  drawRowBar = FALSE,
  titleFontSize = 1.3,
  gene_mar = 10
)

dev.off()
